+++ scripts/scratch.js
@@ -47,6 +47,7 @@ item.system.prof.hasProficiency
 		speaker: ChatMessage.getSpeaker({ actor: actor })});
 
 
+	const damageTypes = [['🧪 Acid', 'acid'], ['❄️ Cold', 'cold'], ['🔥 Fire', 'fire'], ['⚡ Lightning', 'lightning'], ['☁️ Thunder', 'thunder']]; //All possible damage types
 
 	function isAvailableThisTurn() {
 	if (game.combat) {
@@ -188,3 +189,13 @@ async function wait(ms) { return new Promise(resolve => { setTimeout(resolve, ms
 			return !missingDep;
 		}
 
+
+// let rows = "";
+// _steedOptions.forEach(item => {
+// 	// let row = `<div><input type="checkbox" style="margin-right:10px;"/><img style='display:block; width:100%; height:auto;' src='${item.image}'/><label>${item.name}</label></div>` +
+// 	// 	`<br/><div><label>${item.stats}</label></div>`;
+// 	let row = `<label class="radio-label">
+// 		<input type="radio" name="steed" value="${weapon.id}">
+// 		<img src="${weapon.img}" style="border:0px; width: 50px; height:50px;">${weapon.name}</label>`;
+// 	rows += row;
+// });